yolo_training/requirements.txt
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0
pillow>=9.0.0
numpy>=1.21.0
pyyaml>=6.0
opencv-python>=4.5.0
matplotlib>=3.5.0
seaborn>=0.12.0
pandas>=1.3.0

